test_tifs: split into test_insar and test_rtc
